+++ src/components/film-details-controls.js
@@ -1,11 +1,14 @@
 import {AbstractComponent} from './abstract-component.js';
 
 export default class FilmDetailsControls extends AbstractComponent{
-  constructor(watched, watchlist, favorite) {
+  constructor(watched, watchlist, favorite, onDataChange) {
     super();
     this._watched = watched;
     this._watchlist = watchlist;
     this._favorite = favorite;
+    this._onDataChange = onDataChange;
+
+    this._init();
   }
   
   getTemplate() {
@@ -19,5 +22,22 @@ export default class FilmDetailsControls extends AbstractComponent{
     <input type="checkbox" class="film-details__control-input visually-hidden" id="favorite" name="favorite" ${this._favorite ? `checked` : ``}>
     <label for="favorite" class="film-details__control-label film-details__control-label--favorite">Add to favorites</label>
   </section>`
+  }
+
+  _init() {
+    this._onChangeControls(`watchlist`);
+    this._onChangeControls(`watched`);
+    this._onChangeControls(`favorite`);
+  }
+
+  _onChangeControls(checkbox) {
+    const selector = `.film-details__control-label--${checkbox}`;
+    this.getElement()
+      .querySelector(selector)
+        .addEventListener(`click`, (evt) => {
+          const target = evt.target.getAttribute(`for`);
+          console.log(`click`)
+          this._onDataChange(target);
+        });
   }
 }
